Move user related customization to seperate file
